requirements.txt
matplotlib==3.10.5
numpy==2.3.2
pandas==2.3.2
Pillow==11.3.0
scikit_learn==1.7.1
timm==1.0.19
torch==2.8.0
torchsummary==1.5.1
torchvision==0.23.0
